Fix compiler warning about int/size_t comparison

+++ src/palette/internal/palette/palettecreator.cpp
@@ -1102,7 +1102,7 @@ PalettePanel* PaletteCreator::newBracketsPalettePanel()
     static Staff bracketItemOwner(nullptr);
     bracketItemOwner.setBracketType(types.size() - 1, BracketType::NORMAL);
 
-    for (int i = 0; i < types.size(); ++i) {
+    for (size_t i = 0; i < types.size(); ++i) {
         auto b1 = makeElement<Bracket>(gscore);
         auto bi1 = bracketItemOwner.brackets()[i];
         const auto& type = types[i];
